Move RBS code to bravais_settings module
